backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.9
pytesseract==0.3.13
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-dateutil==2.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2
aiofiles==24.1.0
